@@ -3,19 +3,17 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "axios": "^0.18.0",
-    "jsonwebtoken": "^8.5.1",
-    "lodash": "^4.17.11",
-    "luxon": "^1.12.1",
-    "node-sass": "^4.14.1",
-    "prop-types": "^15.7.2",
-    "qs": "^6.7.0",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "2.1.8",
-    "semiotic": "^1.19.7",
-    "styled-components": "^4.2.0"
+    "axios": "^0.27.2",
+    "lodash": "^4.17.21",
+    "luxon": "^3.4.4",
+    "prop-types": "^15.8.1",
+    "qs": "^6.11.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.19.0",
+    "react-scripts": "5.0.1",
+    "semiotic": "^2.0.1",
+    "styled-components": "^6.1.1"
   },
   "scripts": {
     "start": "react-scripts start",
