@@ -2,9 +2,11 @@ import React, { Component } from 'react';
 import api from '../api/routes';
 import Loading from "../components/common/loading";
 import DeviceItem from "../components/device/device-item";
-import DateTime from "luxon/src/datetime";
+import {DateTime} from "luxon";
 import {formatGraphData} from "../utils/graph-data-formatter";
-
+import {
+  useParams,
+} from "react-router-dom"
 
 class Device extends Component {
   constructor(props) {
@@ -24,12 +26,12 @@ class Device extends Component {
     };
   }
 
-  componentWillMount() {
-    this.fetchData(this.props.match.params.id, this.state.startTime, this.state.endTime);
+  componentDidMount() {
+    this.fetchData(this.props.params.id, this.state.startTime, this.state.endTime);
   }
 
-  componentWillReceiveProps(newProps) {
-    if(newProps.match.params.id !== this.props.match.params.id) {
+  componentDidUpdate(newProps) {
+    if(newProps.params.id !== this.props.params.id) {
       this.setState({
         graphData: [],
         loadedDays: 0,
@@ -39,7 +41,7 @@ class Device extends Component {
       const endTime = (new Date()).toISOString();
       const startTime = DateTime.fromISO(endTime).minus({days: 2}).toISO();
 
-      this.fetchData(newProps.match.params.id, startTime, endTime);
+      this.fetchData(newProps.params.id, startTime, endTime);
     }
   }
 
@@ -69,7 +71,7 @@ class Device extends Component {
     const newStartTime = DateTime.fromISO(this.state.startTime).minus({days: 2}).toISO();
     const newEndTime = DateTime.fromISO(this.state.startTime).minus({milliseconds: 1}).toISO();
 
-    return this.fetchData(this.props.match.params.id, newStartTime, newEndTime);
+    return this.fetchData(this.props.params.id, newStartTime, newEndTime);
   }
 
   render() {
@@ -90,6 +92,15 @@ class Device extends Component {
       </div>
     );
   }
+};
+
+const DeviceWithRouter = () => {
+  let params = useParams();
+  return (
+    <Device
+      params={params}
+    />
+  );
 }
 
-export default Device;
+export default DeviceWithRouter;
