@@ -1,5 +1,5 @@
 import React from 'react';
-import styled from 'styled-components/macro';
+import styled from 'styled-components';
 import PropTypes from "prop-types";
 import DeviceLogo from "../deviceSummary/device-logo";
 import DeviceSummaryTemperature from "../deviceSummary/device-summary-temperature";
