@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import styled from 'styled-components/macro';
+import styled from 'styled-components';
 import loadingIcon from "../../assets/icons/loading.svg";
 
 export const StyledLoading= styled.div`
